+++ actions/update-profile.js
@@ -1,6 +1,6 @@
-import darkMode from 'dark-mode';
 import {config} from '../config.js';
 import {
+	getCurrentMode,
 	isAutomaticSwitchingEnabled,
 	isTerminalOpen,
 	setTerminalProfile,
@@ -19,7 +19,7 @@ export async function updateProfile() {
 		throw new Error('Light profile not set');
 	}
 
-	const mode = (await darkMode.isEnabled()) ? 'dark' : 'light';
+	const mode = await getCurrentMode();
 	const profile = config[`${mode}Profile`];
 
 	await setTerminalProfile(profile);